Balance screen

+++ app/src/main/java/ya/co/yandex_finance/component/fragment/balance/BalanceView.kt
@@ -3,5 +3,5 @@ package ya.co.yandex_finance.component.fragment.balance
 import com.arellomobile.mvp.MvpView
 
 interface BalanceView: MvpView {
-    fun setBalance(balance: Int)
+    fun setBalance(balance: Double)
 }
